+++ src/main.tsx
@@ -3,14 +3,17 @@ import ReactDOM from "react-dom/client";
 import App from "./App";
 import { CrawlEventsProvider } from "./hooks/useCrawlEvents";
 import { ToastProvider } from "./hooks/useToasts";
+import { ThemeProvider } from "./hooks/useTheme";
 import "./styles/index.css";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
     <ToastProvider>
-      <CrawlEventsProvider>
-        <App />
-      </CrawlEventsProvider>
+      <ThemeProvider>
+        <CrawlEventsProvider>
+          <App />
+        </CrawlEventsProvider>
+      </ThemeProvider>
     </ToastProvider>
   </React.StrictMode>
 );